Crate btparse

An error utility library for deserializing std::backtrace::Backtrace’s based on its Debug format.

Structs

Backtrace

A deserialized Backtrace.

Error

An error that prevented a backtrace from being deserialized.

Frame

A backtrace frame.

Functions

deserialize

Deserialize a backtrace based on its debug format and return a parsed representation containing a vector of frames.